Seperate useUndoStack hook into it's own file
