+++ src/components/Vizmapper/Forms/DefaultValueForm.tsx
@@ -9,7 +9,7 @@ import { IdType } from '../../../models/IdType'
 import { useVisualStyleStore } from '../../../store/VisualStyleStore'
 
 import { VisualPropertyValueForm } from './VisualPropertyValueForm'
-import { useUndoStack } from '../../../task/ApplyVisualStyle'
+import { useUndoStack } from '../../../task/UndoStack'
 import { UndoCommandType } from '../../../models/StoreModel/UndoStoreModel'
 
 export function DefaultValueForm(props: {
